Filter TargetDirAnnotations between Driver.execute (#240)
This removes all TargetDirAnnotations after running
Driver.getAnnotations. This fixes a regression due to Stage/Phase
where the underlying FirrtlStage.

The underlying bug is that Driver.getAnnotations generates a
TargetDirAnnotation derived from internal CommonOptions. This results
in a duplicate TargetDirAnnotation as the CommonOptions.targetDirName
member is again converted to a TargetDirAnnotation by the Driver
compatibility layer around FirrtlStage.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>