Polyfill formerly deprecated now removed APIs
Many APIs that have been deprecated in firrtl and chisel3 for a long
time are removed in anticipation of the X.5 release. This project is now
on life support--users should switch to using chiseltest (aka
chisel-testers2).

There is a limited compatibility layer for iotesters.PeekPokeTesters,
see https://github.com/ucb-bar/chisel-testers2/pull/348.